worker/requirements.txt
opentelemetry-sdk==1.11.1
opentelemetry-instrumentation-requests==0.30b1
opentelemetry-instrumentation-logging==0.30b1
opentelemetry-instrumentation-psycopg2==0.30b1
opentelemetry-instrumentation-pika==0.30b1
opentelemetry-instrumentation-digma==0.5.16
python-dotenv
pika
psycopg2-binary
pydantic